Recognizes SRT/END and Comments
Parser can now recognize SRT or END and judge if they are correctly implemented (to an extent).
Can also recognize comments in a line and ignore them, allowing the parser to only look at labels/opcodes.
Next, need to make it so that if anything comes before SRT or after END, parser throws an error.