Label, Comment, Class Opcode, Class PALParser Changes
Made some changes to how labels work. Instead of requiring a colon at the end of a label in a branch instruction, one can just use the label's name. Comments have been removed from the .pal file entirely based on a review of the assignment requirements; this level of .log detail was unnecessary and unwanted.
Also, fixed some bugs with how labels were handled in branch modules in Class Opcode. Now, if labels are incorrectly used or not used at all, the correct error will be triggered.
Lastly, Class PALParser has a new addition. In regard to labels, now a warning signals in the .log if a label was not used in a branch statement, helping to prevent accidental exclusions.