Forward Jumps Now Detected
Added some functionality to how labels are treated. Instead of judging a branch as it is encountered, branch labels are added to a list & compared to valid labels after the program has been scanned. Then, an error message will appear.
Also, added folder for .pal files and a folder for .log files. Much easier to see where everything is/should be.